backend/requirements.txt
# Core AI/ML Dependencies
strands-ai>=0.1.0
boto3>=1.34.0
botocore>=1.34.0

# AWS Bedrock AgentCore Dependencies
bedrock-agentcore>=1.0.0
bedrock-agentcore-starter-toolkit>=1.0.0

# HTTP and API Dependencies
requests>=2.31.0
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0